@@ -5,15 +5,16 @@
     "start": "NODE_ENV=production node server/server.js"
   },
   "dependencies": {
-    "@turf/transform-rotate": "^4.4.0",
-    "@turf/turf": "^4.4.0",
+    "@turf/transform-rotate": "^5.1.5",
+    "@turf/turf": "^5.1.6",
     "axios": "^0.16.1",
-    "body-parser": "^1.17.1",
-    "compression": "^1.6.2",
-    "d3-geo": "^1.6.4",
-    "express": "^4.14.0",
-    "express-rate-limit": "^2.8.0",
-    "geoip-lite": "^1.2.1",
+    "body-parser": "^1.18.3",
+    "compression": "^1.7.3",
+    "cors": "^2.8.5",
+    "d3-geo": "^1.11.3",
+    "express": "^4.16.4",
+    "express-rate-limit": "^2.14.2",
+    "geoip-lite": "^1.3.6",
     "immutable": "^3.8.1",
     "lodash.debounce": "^4.0.8",
     "lodash.find": "^4.6.0",
@@ -23,15 +24,15 @@
     "lodash.minby": "^4.6.0",
     "lodash.sample": "^4.2.1",
     "lodash.samplesize": "^4.2.0",
-    "mapbox-gl": "^0.37.0",
+    "mapbox-gl": "^0.53.0",
     "mapbox.js": "^3.1.1",
     "mongodb": "^2.2.26",
-    "next": "^3.2.3",
+    "next": "^5.1.0",
     "next-redux-wrapper": "^1.1.2",
-    "prop-types": "^15.5.8",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-ga": "^2.2.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "react-ga": "^2.5.7",
     "react-mapbox-gl": "^1.11.0",
     "react-redux": "^5.0.4",
     "redis": "^2.7.1",
