@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React, { PureComponent } from 'react';
 import { connect } from 'react-redux';
 import Router from 'next/router';
@@ -7,21 +8,22 @@ import MapInfoBox from './components/MapInfoBox';
 import MapActions from './components/MapActions';
 
 import * as COLORS from '../shared/style/colors';
+import { prefixURL } from '../../utils/url';
 
 let Map;
 
 class MapModal extends PureComponent {
 
   static propTypes = {
-    onDismiss: React.PropTypes.func,
-    isVisible: React.PropTypes.bool,
-    citySlug: React.PropTypes.string,
-    cityName: React.PropTypes.string,
-    activeVehicle: React.PropTypes.string,
-    ownGuess: React.PropTypes.object,
-    itemId: React.PropTypes.number,
-    areaType: React.PropTypes.string,
-    isFetchingLaneData: React.PropTypes.bool
+    onDismiss: PropTypes.func,
+    isVisible: PropTypes.bool,
+    citySlug: PropTypes.string,
+    cityName: PropTypes.string,
+    activeVehicle: PropTypes.string,
+    ownGuess: PropTypes.object,
+    itemId: PropTypes.number,
+    areaType: PropTypes.string,
+    isFetchingLaneData: PropTypes.bool
   }
 
   constructor (props) {
@@ -54,7 +56,7 @@ class MapModal extends PureComponent {
           vehicleType: this.props.activeVehicle
         }
       },{
-        pathname: `/${this.props.citySlug}/explore/${this.props.activeVehicle}`,
+        pathname: prefixURL(`/${this.props.citySlug}/explore/${this.props.activeVehicle}`),
         query: this.props.ownGuess.toJS()
       });
     }
@@ -80,7 +82,7 @@ class MapModal extends PureComponent {
           cityName={this.props.cityName}
         />
         <div className="BtnCloseMap" onClick={this.closeModal}>
-          <img src="/static/icons/Icon_Cross.svg" alt="close map"/>
+          <img src={prefixURL("/static/icons/Icon_Cross.svg")} alt="close map"/>
         </div>
         <div className={`LoaderWrapper ${!this.state.mapLoaded || this.props.isFetchingLaneData ? 'visible' : 'hidden'}`}>
           <Loader />
