@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React, { Component } from "react";
 import ReactMapboxGl, { ScaleControl, ZoomControl } from "react-mapbox-gl";
 import TWEEN from "tween.js";
@@ -5,6 +6,10 @@ import { bbox } from "@turf/turf";
 import rotate from "@turf/transform-rotate";
 import Config from "../../../config.json";
 import * as d3geo from "d3-geo";
+
+import getConfig from 'next/config'
+const { publicRuntimeConfig } = getConfig()
+const { env_mapbox_token } = publicRuntimeConfig
 
 import MapActions from "./MapActions";
 
@@ -26,11 +31,11 @@ const unfolder = unfold();
 
 class Map extends Component {
   static propTypes = {
-    activeVehicle: React.PropTypes.string,
-    areaType: React.PropTypes.string,
-    laneData: React.PropTypes.object,
-    parkingData: React.PropTypes.object,
-    onMapLoaded: React.PropTypes.func
+    activeVehicle: PropTypes.string,
+    areaType: PropTypes.string,
+    laneData: PropTypes.object,
+    parkingData: PropTypes.object,
+    onMapLoaded: PropTypes.func
   };
 
   constructor(props) {
