@@ -1,18 +1,19 @@
+import PropTypes from 'prop-types';
 import React from 'react';
-
-// Images
-const ArrowActualData = '/static/icons/Arrow_ActualData.svg';
-
 import * as METRICS from '../../shared/style/metrics';
 import * as COLORS from '../../shared/style/colors';
+import { prefixURL } from '../../../utils/url';
+
+// Images
+const ArrowActualData = prefixURL('/static/icons/Arrow_ActualData.svg');
 
 class ActualValueArrow extends React.PureComponent {
 
   static propTypes = {
-    data: React.PropTypes.number,
-    city: React.PropTypes.string,
-    position: React.PropTypes.string,
-    offset: React.PropTypes.number,
+    data: PropTypes.number,
+    city: PropTypes.string,
+    position: PropTypes.string,
+    offset: PropTypes.number,
   }
 
   renderAbove = () => (
