@@ -1,14 +1,16 @@
+import PropTypes from 'prop-types';
 import React from 'react';
+import { prefixURL } from '../../../utils/url';
 
 // Images
-const ArrowOtherGuess = '/static/icons/Arrow_OtherCity.svg';
+const ArrowOtherGuess = prefixURL('/static/icons/Arrow_OtherCity.svg');
 
 class OthersGuessesArrow extends React.PureComponent {
 
   static propTypes = {
-    data: React.PropTypes.array,
-    position: React.PropTypes.string,
-    offset: React.PropTypes.number,
+    data: PropTypes.array,
+    position: PropTypes.string,
+    offset: PropTypes.number,
   }
 
   getSortedData = (data) => data.sort((a, b) => {
