@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import Router from 'next/router';
 import _find from 'lodash.find';
@@ -10,12 +11,13 @@ import GifItem from '../components/GifItem';
 import RoundedButton from '../../shared/components/RoundedButton';
 
 import GifData from '../../../gifgallery.json';
+import { getRootURL } from '../../../utils/url';
 
 class GifSection extends React.PureComponent {
 
   static propTypes = {
-    city: React.PropTypes.object,
-    ownGuess: React.PropTypes.object
+    city: PropTypes.object,
+    ownGuess: PropTypes.object
   }
   
   constructor(props) {
@@ -40,7 +42,7 @@ class GifSection extends React.PureComponent {
   }
 
   getShareUrl(staffPickData) {
-   return `https://whatthestreet.moovellab.com/${staffPickData.citySlug}/explore/car/lanes/${staffPickData.id}`; 
+   return `https://${getRootURL()}/${staffPickData.citySlug}/explore/car/lanes/${staffPickData.id}`; 
   }
 
   getShareText(staffPickData) {
@@ -117,7 +119,7 @@ class GifSection extends React.PureComponent {
                       pathname: '/explore',
                       query: this.props.ownGuess
                     },{
-                      pathname: `/${this.props.city.slug}/explore`,
+                      pathname: prefixURL(`/${this.props.city.slug}/explore`),
                       query: this.props.ownGuess
                     });
                   }}
