@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
@@ -8,22 +9,25 @@ import ResultsBarChart from '../components/ResultsBarChart';
 import ComparisonBarChart from '../components/ComparisonBarChart';
 import DeviderImage from '../components/DeviderImage';
 
-// Images
-const SnapDeviderBackground = '/static/images/Background_Highway.jpg';
-const IllustrationSnap = '/static/icons/Illustration_Snap.svg';
-
 // Selectors
 import { CitySelectors, GuessSelectors } from '../../statemanagement/selectors';
+import { prefixURL } from '../../../utils/url';
+
+// Images
+const SnapDeviderBackground = prefixURL('/static/images/Background_Highway.jpg');
+const IllustrationSnap = prefixURL('/static/icons/Illustration_Snap.svg');
+
+
 
 class ResultsPage extends React.PureComponent {
 
   static propTypes = {
-    own: React.PropTypes.objectOf(React.PropTypes.number),
-    others: React.PropTypes.arrayOf(React.PropTypes.object),
-    actual: React.PropTypes.object,
-    city: React.PropTypes.object,
-    suggestion: React.PropTypes.object,
-    isLoading: React.PropTypes.bool,
+    own: PropTypes.objectOf(PropTypes.number),
+    others: PropTypes.arrayOf(PropTypes.object),
+    actual: PropTypes.object,
+    city: PropTypes.object,
+    suggestion: PropTypes.object,
+    isLoading: PropTypes.bool,
   }
 
   constructor(props) {
